Move slugger to external file
