+++ lib/marked.js
@@ -1268,33 +1268,6 @@ Parser.prototype.tok = function() {
 };
 
 /**
- * Slugger
- */
-function Slugger () {
-  this.seen = {};
-}
-
-/**
- * Generate a unique slug.
- */
-Slugger.prototype.slug = function (value) {
-  var slug = value
-    .toLowerCase()
-    .trim()
-    .replace(/[\u2000-\u206F\u2E00-\u2E7F\\'!"#$%&()*+,./:;<=>?@[\]^`{|}~]/g, '')
-    .replace(/\s/g, '-');
-
-  var count = this.seen.hasOwnProperty(slug) ? this.seen[slug] + 1 : 0;
-  this.seen[slug] = count;
-
-  if (count > 0) {
-    slug = slug + '-' + count;
-  }
-
-  return slug;
-};
-
-/**
  * Helpers
  */
 
@@ -1626,8 +1599,6 @@ marked.lexer = Lexer.lex;
 
 marked.InlineLexer = InlineLexer;
 marked.inlineLexer = InlineLexer.output;
-
-marked.Slugger = Slugger;
 
 marked.parse = marked;
